fix: get actual user id from database
The post validation function compares two id numbers to check for owner
of post. They are taken from the authorization token, and the user id
from the database, searched by using the email.
The endpoint used the post id, which can sometimes coincide with the
user id. Hence, the error.